refactor(ICache): move SRAMTemplateWithFixedWidth into seperate file
