+++ src/main/scala/xiangshan/frontend/icache/ICache.scala
@@ -755,136 +755,3 @@ class ICacheImp(outer: ICache) extends LazyModuleImp(outer) with HasICacheParame
   )
   generatePerfEvent()
 }
-
-//class ICachePartWayReadBundle[T <: Data](gen: T, pWay: Int)(implicit p: Parameters)
-//    extends ICacheBundle {
-//  val req = Flipped(Vec(
-//    PortNumber,
-//    Decoupled(new Bundle {
-//      val ridx = UInt((log2Ceil(nSets) - 1).W)
-//    })
-//  ))
-//  val resp = Output(new Bundle {
-//    val rdata = Vec(PortNumber, Vec(pWay, gen))
-//  })
-//}
-
-//class ICacheWriteBundle[T <: Data](gen: T, pWay: Int)(implicit p: Parameters)
-//    extends ICacheBundle {
-//  val wdata    = gen
-//  val widx     = UInt((log2Ceil(nSets) - 1).W)
-//  val wbankidx = Bool()
-//  val wmask    = Vec(pWay, Bool())
-//}
-
-//class ICachePartWayArray[T <: Data](gen: T, pWay: Int)(implicit p: Parameters) extends ICacheArray {
-//
-//  // including part way data
-//  val io = IO {
-//    new Bundle {
-//      val read  = new ICachePartWayReadBundle(gen, pWay)
-//      val write = Flipped(ValidIO(new ICacheWriteBundle(gen, pWay)))
-//    }
-//  }
-//
-//  io.read.req.map(_.ready := !io.write.valid)
-//
-//  val srams = (0 until PortNumber) map { bank =>
-//    val sramBank = Module(new SRAMTemplate(
-//      gen,
-//      set = nSets / 2,
-//      way = pWay,
-//      shouldReset = true,
-//      holdRead = true,
-//      singlePort = true,
-//      withClockGate = true
-//    ))
-//
-//    sramBank.io.r.req.valid := io.read.req(bank).valid
-//    sramBank.io.r.req.bits.apply(setIdx = io.read.req(bank).bits.ridx)
-//
-//    if (bank == 0) sramBank.io.w.req.valid := io.write.valid && !io.write.bits.wbankidx
-//    else sramBank.io.w.req.valid           := io.write.valid && io.write.bits.wbankidx
-//    sramBank.io.w.req.bits.apply(
-//      data = io.write.bits.wdata,
-//      setIdx = io.write.bits.widx,
-//      waymask = io.write.bits.wmask.asUInt
-//    )
-//
-//    sramBank
-//  }
-//
-//  io.read.req.map(_.ready := !io.write.valid && srams.map(_.io.r.req.ready).reduce(_ && _))
-//
-//  io.read.resp.rdata := VecInit(srams.map(bank => bank.io.r.resp.asTypeOf(Vec(pWay, gen))))
-//
-//}
-
-class SRAMTemplateWithFixedWidthIO[T <: Data](gen: T, set: Int, way: Int) extends Bundle {
-  val r: SRAMReadBus[T]  = Flipped(new SRAMReadBus(gen, set, way))
-  val w: SRAMWriteBus[T] = Flipped(new SRAMWriteBus(gen, set, way))
-}
-
-// Automatically partition the SRAM based on the width of the data and the desired width.
-// final SRAM width = width * way
-class SRAMTemplateWithFixedWidth[T <: Data](
-    gen:           T,
-    set:           Int,
-    width:         Int,
-    way:           Int = 1,
-    shouldReset:   Boolean = false,
-    holdRead:      Boolean = false,
-    singlePort:    Boolean = false,
-    bypassWrite:   Boolean = false,
-    withClockGate: Boolean = false,
-    hasMbist:      Boolean = false
-) extends Module {
-
-  private val dataBits  = gen.getWidth
-  private val bankNum   = math.ceil(dataBits.toDouble / width.toDouble).toInt
-  private val totalBits = bankNum * width
-
-  val io: SRAMTemplateWithFixedWidthIO[T] = IO(new SRAMTemplateWithFixedWidthIO(gen, set, way))
-
-  private val wordType = UInt(width.W)
-  private val writeDatas = (0 until bankNum).map { bank =>
-    VecInit((0 until way).map { i =>
-      io.w.req.bits.data(i).asTypeOf(UInt(totalBits.W)).asTypeOf(Vec(bankNum, wordType))(bank)
-    })
-  }
-
-  private val srams = (0 until bankNum) map { bank =>
-    val sramBank = Module(new SRAMTemplate(
-      wordType,
-      set = set,
-      way = way,
-      shouldReset = shouldReset,
-      holdRead = holdRead,
-      singlePort = singlePort,
-      bypassWrite = bypassWrite,
-      withClockGate = withClockGate,
-      hasMbist = hasMbist
-    ))
-    // read req
-    sramBank.io.r.req.valid       := io.r.req.valid
-    sramBank.io.r.req.bits.setIdx := io.r.req.bits.setIdx
-
-    // write req
-    sramBank.io.w.req.valid       := io.w.req.valid
-    sramBank.io.w.req.bits.setIdx := io.w.req.bits.setIdx
-    sramBank.io.w.req.bits.data   := writeDatas(bank)
-    sramBank.io.w.req.bits.waymask.foreach(_ := io.w.req.bits.waymask.get)
-
-    sramBank
-  }
-
-  io.r.req.ready := !io.w.req.valid
-  (0 until way).foreach { i =>
-    io.r.resp.data(i) := VecInit((0 until bankNum).map(bank =>
-      srams(bank).io.r.resp.data(i)
-    )).asTypeOf(UInt(totalBits.W))(dataBits - 1, 0).asTypeOf(gen.cloneType)
-  }
-
-  io.r.req.ready := srams.head.io.r.req.ready
-  io.w.req.ready := srams.head.io.w.req.ready
-}
